requirements.txt
flask
line-bot-sdk
pymongo[tls]
requests
flask-wtf
wtforms
email_validator
flask_bcrypt
flask_login



